Multi-format output and parametric demo
* Add a PHP entry point which takes an associative array of options, and
  produces diffs in multiple formats simultaneously.
* Allow option configuration in demo.php.
* Update README.md.

+++ src/lib/LineDiffProcessor.h
@@ -46,7 +46,7 @@ class LineDiffProcessor {
 			 * The maximum number of RHS lines which can be compared with
 			 * one LHS line.
 			 */
-			int maxSplitSize;
+			int64_t maxSplitSize;
 		};
 
 		LineDiffProcessor(const Config & config_, WordDiffCache & wordDiffCache_)
