+++ src/lib/Wikidiff2.h
@@ -65,7 +65,7 @@ class Wikidiff2 {
 			 * The maximum number of RHS lines which can be compared with
 			 * one LHS line.
 			 */
-			int maxSplitSize;
+			int64_t maxSplitSize;
 
 			/**
 			 * The minimum similarity which must be maintained during a split
